various things, branch dj/188 (#275)
ENCLONE VISUAL
- Improve logging.
- Dynamically determine wait time for timeout.
- Prevent line wrapping.
- Improve test organization and add test.
- Add Del button to delete current state from history.
- Fix bug.
- Keep trying to connect for up to ten seconds.
OTHER
- The install script now records the size parameter in a file ~/enclone/size.
- Despecify crates and update them.
- Add a program to update crates if compilation succeeds.
STATE
- ./test, enclone.test and test_vis all pass; also manual tests were run after all changes that could affect them

+++ enclone_visual/src/bin/test_vis.rs
@@ -10,6 +10,13 @@
 // having larger files in git.  A better solution would be to use lowest resolution
 // JPEG2000 files, which would be even smaller.
 //
+// You need the following datasets to run this:
+// dataset   notes
+// 123085    public
+// 123217    public, but we use our internal copy, which includes feature_barcode_matrix.bin
+// 1145040   not public
+// 1142282   not public
+//
 // See also show_diffs.
 
 use enclone_visual::compare_images::*;
@@ -99,8 +106,10 @@ fn main() {
         let new_jpg_file = format!("{}.jpg", new_png_file.rev_before(".png"));
         {
             let quality = 1 as u8; // lowest quality
-                                   // This file removal is to circumvent a bug in Carbon Black.
-            std::fs::remove_file(&new_jpg_file).unwrap();
+            if path_exists(&new_jpg_file) {
+                // This file removal is to circumvent a bug in Carbon Black.
+                std::fs::remove_file(&new_jpg_file).unwrap();
+            }
             let mut f = open_for_write_new![&new_jpg_file];
             let mut buff = BufWriter::new(&mut f);
             let mut encoder = JpegEncoder::new_with_quality(&mut buff, quality);
@@ -158,7 +167,7 @@ fn main() {
     // ▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓
 
     let used = elapsed(&t);
-    const EXPECTED_TIME: f64 = 23.2; // this is supposed to be the lowest observed value
+    const EXPECTED_TIME: f64 = 25.1; // this is supposed to be the lowest observed value
     const MAX_PERCENT_OVER: f64 = 4.0;
     let percent_over = 100.0 * (used - EXPECTED_TIME) / EXPECTED_TIME;
     if percent_over > MAX_PERCENT_OVER {
@@ -266,8 +275,8 @@ fn main() {
     // ▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓
 
     println!(
-        "\nenclone visual tests completely {} in {:.1} seconds using {:.1} MB\n",
-        state, used, peak_mem_mb,
+        "\nenclone visual tests completed {} in {:.1} seconds",
+        state, used,
     );
     println!("actual total time = {:.1} seconds\n", elapsed(&tall));
     if fail {
